Splitting up msgpack.ts into one class per file.  Added test generously provided by @dOrgJelli.
