+++ assembly/__tests__/as-pect.d.ts
@@ -1,0 +1,1 @@
+/// <reference path="../../node_modules/@as-pect/assembly/types/as-pect.d.ts" />
