+++ assembly/decoder.ts
@@ -1,4 +1,5 @@
 import { DataReader } from "./datareader";
+import { Format } from "./format";
 import { E_INVALIDLENGTH } from "util/error";
 
 export class Decoder {
@@ -57,8 +58,11 @@ export class Decoder {
   readInt64(): i64 {
     const prefix = this.reader.getUint8();
 
-    if (this.isFixedInt(prefix) || this.isNegativeFixedInt(prefix)) {
+    if (this.isFixedInt(prefix)) {
       return <i64>prefix;
+    }
+    if (this.isNegativeFixedInt(prefix)) {
+      return <i64>(<i8>prefix);
     }
     switch (prefix) {
       case Format.INT8:
@@ -418,420 +422,4 @@ export class Decoder {
     }
     return this.readMap(keyFn, valueFn);
   }
-}
-
-export class MapEntry<K, V> {
-  key: K;
-  value: V;
-
-  constructor(key: K, value: V) {
-    this.key = key;
-    this.value = value;
-  }
-}
-
-export class Encoder {
-  private reader: DataReader;
-
-  constructor(ua: ArrayBuffer) {
-    this.reader = new DataReader(ua, 0, ua.byteLength);
-  }
-
-  writeNil(): void {
-    this.reader.setUint8(<u8>Format.NIL);
-  }
-
-  writeBool(value: bool): void {
-    this.reader.setUint8(value ? <u8>Format.TRUE : <u8>Format.FALSE);
-  }
-
-  writeInt8(value: i8): void {
-    this.writeInt64(<i64>value);
-  }
-  writeInt16(value: i16): void {
-    this.writeInt64(<i64>value);
-  }
-  writeInt32(value: i32): void {
-    this.writeInt64(<i64>value);
-  }
-  writeInt64(value: i64): void {
-    if (value >= 0 && value < 1 << 7) {
-      this.reader.setUint8(<u8>value);
-    } else if (value < 0 && value >= -(1 << 5)) {
-      this.reader.setUint8((<u8>value) | (<u8>Format.NEGATIVE_FIXINT));
-    } else if (value <= <i64>i8.MAX_VALUE && value >= <i64>i8.MIN_VALUE) {
-      this.reader.setUint8(<u8>Format.INT8);
-      this.reader.setInt8(<i8>value);
-    } else if (value <= <i64>i16.MAX_VALUE && value >= <i64>i16.MIN_VALUE) {
-      this.reader.setUint8(<u8>Format.INT16);
-      this.reader.setInt16(<i16>value);
-    } else if (value <= <i64>i32.MAX_VALUE && value >= <i64>i32.MIN_VALUE) {
-      this.reader.setUint8(<u8>Format.INT32);
-      this.reader.setInt32(<i32>value);
-    } else {
-      this.reader.setUint8(<u8>Format.INT64);
-      this.reader.setInt64(value);
-    }
-  }
-
-  writeUInt8(value: u8): void {
-    this.writeUInt64(<u64>value);
-  }
-  writeUInt16(value: u16): void {
-    this.writeUInt64(<u64>value);
-  }
-  writeUInt32(value: u32): void {
-    this.writeUInt64(<u64>value);
-  }
-  writeUInt64(value: u64): void {
-    if (value < 1 << 7) {
-      this.reader.setUint8(<u8>value);
-    } else if (value <= <u64>u8.MAX_VALUE) {
-      this.reader.setUint8(<u8>Format.UINT8);
-      this.reader.setUint8(<u8>value);
-    } else if (value <= <u64>u16.MAX_VALUE) {
-      this.reader.setUint8(<u8>Format.UINT16);
-      this.reader.setUint16(<u16>value);
-    } else if (value <= <u64>u32.MAX_VALUE) {
-      this.reader.setUint8(<u8>Format.UINT32);
-      this.reader.setUint32(<u32>value);
-    } else {
-      this.reader.setUint8(<u8>Format.UINT64);
-      this.reader.setUint64(value);
-    }
-  }
-
-  writeFloat32(value: f32): void {
-    this.reader.setUint8(<u8>Format.FLOAT32);
-    this.reader.setFloat32(value);
-  }
-
-  writeFloat64(value: f64): void {
-    this.reader.setUint8(<u8>Format.FLOAT64);
-    this.reader.setFloat64(value);
-  }
-
-  writeStringLength(length: u32): void {
-    if (length < 32) {
-      this.reader.setUint8((<u8>length) | (<u8>Format.FIXSTR));
-    } else if (length <= <u32>u8.MAX_VALUE) {
-      this.reader.setUint8(<u8>Format.STR8);
-      this.reader.setUint8(<u8>length);
-    } else if (length <= <u32>u16.MAX_VALUE) {
-      this.reader.setUint8(<u8>Format.STR16);
-      this.reader.setUint16(<u16>length);
-    } else {
-      this.reader.setUint8(<u8>Format.STR32);
-      this.reader.setUint32(length);
-    }
-  }
-
-  writeString(value: string): void {
-    const buf = String.UTF8.encode(value);
-    this.writeStringLength(buf.byteLength);
-    this.reader.setBytes(buf);
-  }
-
-  writeBinLength(length: u32): void {
-    if (length <= <u32>u8.MAX_VALUE) {
-      this.reader.setUint8(<u8>Format.BIN8);
-      this.reader.setUint8(<u8>length);
-    } else if (length <= <u32>u16.MAX_VALUE) {
-      this.reader.setUint8(<u8>Format.BIN16);
-      this.reader.setUint16(<u16>length);
-    } else {
-      this.reader.setUint8(<u8>Format.BIN32);
-      this.reader.setUint32(length);
-    }
-  }
-
-  writeByteArray(ab: ArrayBuffer): void {
-    if (ab.byteLength == 0) {
-      this.writeNil();
-      return;
-    }
-    this.writeBinLength(ab.byteLength);
-    this.reader.setBytes(ab);
-  }
-
-  writeArraySize(length: u32): void {
-    if (length < 16) {
-      this.reader.setInt8((<u8>length) | (<u8>Format.FIXARRAY));
-    } else if (length <= <u32>u16.MAX_VALUE) {
-      this.reader.setUint8(<u8>Format.ARRAY16);
-      this.reader.setUint16(<u16>length);
-    } else {
-      this.reader.setUint8(<u8>Format.ARRAY32);
-      this.reader.setUint32(length);
-    }
-  }
-
-  writeMapSize(length: u32): void {
-    if (length < 16) {
-      this.reader.setInt8((<u8>length) | (<u8>Format.FIXMAP));
-    } else if (length <= <u32>u16.MAX_VALUE) {
-      this.reader.setUint8(<u8>Format.MAP16);
-      this.reader.setUint16(<u16>length);
-    } else {
-      this.reader.setUint8(<u8>Format.MAP32);
-      this.reader.setUint32(length);
-    }
-  }
-
-  writeArray<T>(a: Array<T>, fn: (encoder: Encoder, item: T) => void): void {
-    this.writeArraySize(a.length);
-    for (let i: i32 = 0; i < a.length; i++) {
-      fn(this, a[i]);
-    }
-  }
-
-  writeNullableArray<T>(
-    a: Array<T> | null,
-    fn: (encoder: Encoder, item: T) => void
-  ): void {
-    if (a === null) {
-      this.writeNil();
-      return;
-    }
-    this.writeArray(a, fn);
-  }
-
-  writeMap<K, V>(
-    m: Map<K, V>,
-    keyFn: (encoder: Encoder, key: K) => void,
-    valueFn: (encoder: Encoder, value: V) => void
-  ): void {
-    this.writeMapSize(m.size);
-    const keys = m.keys();
-    for (let i: i32 = 0; i < keys.length; i++) {
-      const key = keys[i];
-      const value = m.get(key);
-      keyFn(this, key);
-      valueFn(this, value);
-    }
-  }
-
-  writeNullableMap<K, V>(
-    m: Map<K, V> | null,
-    keyFn: (encoder: Encoder, key: K) => void,
-    valueFn: (encoder: Encoder, value: V) => void
-  ): void {
-    if (m === null) {
-      this.writeNil();
-      return;
-    }
-    this.writeMap(m, keyFn, valueFn);
-  }
-}
-
-export class Sizer {
-  length: i32;
-
-  constructor() {}
-
-  writeNil(): void {
-    this.length++;
-  }
-
-  writeString(value: string): void {
-    const buf = String.UTF8.encode(value);
-    this.writeStringLength(buf.byteLength);
-    this.length += buf.byteLength;
-  }
-
-  writeStringLength(length: u32): void {
-    if (length < 32) {
-      this.length++;
-    } else if (length <= <u32>u8.MAX_VALUE) {
-      this.length += 2;
-    } else if (length <= <u32>u16.MAX_VALUE) {
-      this.length += 3;
-    } else {
-      this.length += 5;
-    }
-  }
-
-  writeBool(value: bool): void {
-    this.length++;
-  }
-
-  writeArraySize(length: u32): void {
-    if (length < 16) {
-      this.length++;
-    } else if (length <= <u32>u16.MAX_VALUE) {
-      this.length += 3;
-    } else {
-      this.length += 5;
-    }
-  }
-
-  writeBinLength(length: u32): void {
-    if (length <= <u32>u8.MAX_VALUE) {
-      this.length += 2;
-    } else if (length <= <u32>u16.MAX_VALUE) {
-      this.length += 3;
-    } else {
-      this.length += 5;
-    }
-  }
-
-  writeByteArray(ab: ArrayBuffer): void {
-    if (ab.byteLength == 0) {
-      this.length++; //nil byte
-      return;
-    }
-    this.writeBinLength(ab.byteLength);
-    this.length += ab.byteLength + 1;
-  }
-
-  writeMapSize(length: u32): void {
-    if (length < 16) {
-      this.length++;
-    } else if (length <= <u32>u16.MAX_VALUE) {
-      this.length += 3;
-    } else {
-      this.length += 5;
-    }
-  }
-
-  writeInt8(value: i8): void {
-    this.writeInt64(<i64>value);
-  }
-  writeInt16(value: i16): void {
-    this.writeInt64(<i64>value);
-  }
-  writeInt32(value: i32): void {
-    this.writeInt64(<i64>value);
-  }
-  writeInt64(value: i64): void {
-    if (value >= -(1 << 5) && value < 1 << 7) {
-      this.length++;
-    } else if (value < 1 << 7 && value >= -(1 << 7)) {
-      this.length += 2;
-    } else if (value < 1 << 15 && value >= -(1 << 15)) {
-      this.length += 3;
-    } else if (value < 1 << 31 && value >= -(1 << 31)) {
-      this.length += 5;
-    } else {
-      this.length += 9;
-    }
-  }
-
-  writeUInt8(value: u8): void {
-    this.writeUInt64(<u64>value);
-  }
-  writeUInt16(value: u16): void {
-    this.writeUInt64(<u64>value);
-  }
-  writeUInt32(value: u32): void {
-    this.writeUInt64(<u64>value);
-  }
-  writeUInt64(value: u64): void {
-    if (value < 1 << 7) {
-      this.length++;
-    } else if (value < 1 << 8) {
-      this.length += 2;
-    } else if (value < 1 << 16) {
-      this.length += 3;
-    } else if (value < 1 << 32) {
-      this.length += 5;
-    } else {
-      this.length += 9;
-    }
-  }
-
-  writeFloat32(value: f32): void {
-    this.length += 5;
-  }
-  writeFloat64(value: f64): void {
-    this.length += 9;
-  }
-
-  writeArray<T>(a: Array<T>, fn: (sizer: Sizer, item: T) => void): void {
-    this.writeArraySize(a.length);
-    for (let i: i32 = 0; i < a.length; i++) {
-      fn(this, a[i]);
-    }
-  }
-
-  writeNullableArray<T>(
-    a: Array<T> | null,
-    fn: (sizer: Sizer, item: T) => void
-  ): void {
-    if (a === null) {
-      this.writeNil();
-      return;
-    }
-    this.writeArray(a, fn);
-  }
-
-  writeMap<K, V>(
-    m: Map<K, V>,
-    keyFn: (sizer: Sizer, key: K) => void,
-    valueFn: (sizer: Sizer, value: V) => void
-  ): void {
-    this.writeMapSize(m.size);
-    const keys = m.keys();
-    for (let i: i32 = 0; i < keys.length; i++) {
-      const key = keys[i];
-      const value = m.get(key);
-      keyFn(this, key);
-      valueFn(this, value);
-    }
-  }
-
-  writeNullableMap<K, V>(
-    m: Map<K, V> | null,
-    keyFn: (sizer: Sizer, key: K) => void,
-    valueFn: (sizer: Sizer, value: V) => void
-  ): void {
-    if (m === null) {
-      this.writeNil();
-      return;
-    }
-    this.writeMap(m, keyFn, valueFn);
-  }
-}
-
-const enum Format {
-  ERROR = 0,
-  Four_Bytes = 0xffffffff,
-  FOUR_LEAST_SIG_BITS_IN_BYTE = 0x0f,
-  FOUR_SIG_BITS_IN_BYTE = 0xf0,
-  POSITIVE_FIXINT = 0x00,
-  FIXMAP = 0x80,
-  FIXARRAY = 0x90,
-  FIXSTR = 0xa0,
-  NIL = 0xc0,
-  FALSE = 0xc2,
-  TRUE = 0xc3,
-  BIN8 = 0xc4,
-  BIN16 = 0xc5,
-  BIN32 = 0xc6,
-  EXT8 = 0xc7,
-  EXT16 = 0xc8,
-  EXT32 = 0xc9,
-  FLOAT32 = 0xca,
-  FLOAT64 = 0xcb,
-  UINT8 = 0xcc,
-  UINT16 = 0xcd,
-  UINT32 = 0xce,
-  UINT64 = 0xcf,
-  INT8 = 0xd0,
-  INT16 = 0xd1,
-  INT32 = 0xd2,
-  INT64 = 0xd3,
-  FIXEXT1 = 0xd4,
-  FIXEXT2 = 0xd5,
-  FIXEXT4 = 0xd6,
-  FIXEXT8 = 0xd7,
-  FIXEXT16 = 0xd8,
-  STR8 = 0xd9,
-  STR16 = 0xda,
-  STR32 = 0xdb,
-  ARRAY16 = 0xdc,
-  ARRAY32 = 0xdd,
-  MAP16 = 0xde,
-  MAP32 = 0xdf,
-  NEGATIVE_FIXINT = 0xe0,
 }
