+++ package.json
@@ -7,9 +7,14 @@
     "type": "git",
     "url": "https://github.com/wapc/as-msgpack.git"
   },
+  "scripts": {
+    "test": "asp --verbose",
+    "test:ci": "asp --summary"
+  },
   "author": "",
   "license": "Apache-2.0",
   "devDependencies": {
-    "assemblyscript": "^0.9.4"
+    "@as-pect/cli": "4.0.0",
+    "assemblyscript": "^0.16.1"
   }
 }
