+++ src/index.d.ts
@@ -1,3 +1,6 @@
-export * from "./msgpack";
+export * from "./decoder";
+export * from "./writer";
+export * from "./encoder";
+export * from "./sizer";
 export * from "./value";
 export * from "./datareader";
